requirements.txt
langgraph>=0.2
langchain-anthropic>=0.2
langgraph-checkpoint-sqlite>=3.0
mcp>=1.0
google-api-python-client>=2.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
playwright>=1.40
fastapi>=0.115
uvicorn>=0.30
python-dotenv>=1.0
rich>=13.0
typer>=0.12
anthropic>=0.39
python-telegram-bot>=21.0
